Move patch code to Hooks.cs and Hooks.LoadMotionList
