+++ src/AnimationLoader.Core/SwapAnim.cs
@@ -1,25 +1,17 @@
 using System;
-using System.Collections;
-using BepInEx;
-using HarmonyLib;
-using IllusionUtility.GetUtility;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Reflection.Emit;
 using System.Xml.Linq;
 using System.Xml.Serialization;
+
+using BepInEx;
 using BepInEx.Configuration;
 using BepInEx.Logging;
-using Illusion.Extensions;
-using Sideloader.AutoResolver;
-using Studio;
-using TMPro;
+
 using UnityEngine;
-using UnityEngine.EventSystems;
-using UnityEngine.UI;
+
 using static HFlag;
-using Manager;
 
 [assembly: System.Reflection.AssemblyFileVersion(AnimationLoader.SwapAnim.Version)]
 
@@ -56,6 +48,8 @@ namespace AnimationLoader
             {"titspussy", "siru_t_khs_n07"},
             {"tits", "siru_t_khh_11"},
             {"pussy", "siru_t_khs_n07"}, // have to make this manually, for now copy TitsPussy
+            {"kksbutt", "siru_t_khs_14" },
+            {"kksfacetits", "siru_t_khh_33" },
         };
 
         private static readonly Dictionary<string, int> EModeGroups = new Dictionary<string, int>
@@ -75,16 +69,23 @@ namespace AnimationLoader
         {
             Logger = base.Logger;
 
-            SortPositions = Config.Bind(GeneralSection, nameof(SortPositions), true, new ConfigDescription("Sort positions alphabetically"));
-            ReloadManifests = Config.Bind(GeneralSection, nameof(ReloadManifests), new KeyboardShortcut(KeyCode.None), new ConfigDescription("Load positions from all manifest format xml files inside config/AnimationLoader folder"));
+            // 
+            SortPositions = Config.Bind(
+                GeneralSection,
+                nameof(SortPositions),
+                true,
+                new ConfigDescription("Sort positions alphabetically"));
+            ReloadManifests = Config.Bind(
+                GeneralSection,
+                nameof(ReloadManifests),
+                new KeyboardShortcut(KeyCode.None),
+                new ConfigDescription("Load positions from all manifest format xml files inside" +
+                " config/AnimationLoader folder"));
+            // Standard game code manages the display of animators buttons no grid UI
+#if KK
             UseGrid = Config.Bind(GeneralSection, nameof(UseGrid), false, new ConfigDescription("If you don't want to use the scrollable list for some reason"));
-            
-            var harmony = Harmony.CreateAndPatchAll(typeof(SwapAnim), nameof(SwapAnim));
-            if(vrType != null)
-            {
-                harmony.Patch(AccessTools.Method(vrType, nameof(HSceneProc.ChangeAnimator)), postfix: new HarmonyMethod(typeof(SwapAnim), nameof(SwapAnimation)));
-                harmony.Patch(AccessTools.Method(vrType, nameof(HSceneProc.CreateAllAnimationList)), postfix: new HarmonyMethod(typeof(SwapAnim), nameof(ExtendList)));
-            }
+#endif
+            Hooks.Init();
         }
 
         private void Start()
@@ -139,239 +140,6 @@ namespace AnimationLoader
             }
         }
 
-        [HarmonyPostfix, HarmonyPatch(typeof(HSceneProc), nameof(HSceneProc.CreateAllAnimationList))]
-        private static void ExtendList(object __instance) {
-            // add new animations to the complete list
-#if KK
-            var hlist = Singleton<Game>.Instance.glSaveData.playHList;
-#elif KKS
-            var hlist = Game.globalData.playHList;
-#endif
-            var lstAnimInfo = Traverse.Create(__instance).Field<List<HSceneProc.AnimationListInfo>[]>("lstAnimInfo").Value;
-            swapAnimationMapping = new Dictionary<HSceneProc.AnimationListInfo, SwapAnimationInfo>();
-            foreach (var anim in animationDict.SelectMany(
-                e => e.Value,
-                (e, a) => a
-            ))
-            {
-                var mode = (int)anim.Mode;
-                if (mode < 0 || mode >= lstAnimInfo.Length)
-                {
-                    continue;
-                }
-                var animListInfo = lstAnimInfo[(int)anim.Mode];
-                var donorInfo = animListInfo.FirstOrDefault(x => x.id == anim.DonorPoseId)?.DeepCopy();
-                if (donorInfo == null)
-                {
-                    Logger.LogWarning($"No donor: {anim.Mode} {anim.DonorPoseId}");
-                    continue;
-                }
-
-                if (anim.NeckDonorId >= 0 && anim.NeckDonorId != anim.DonorPoseId)
-                    donorInfo.paramFemale.fileMotionNeck = animListInfo.First(x => x.id == anim.NeckDonorId).paramFemale.fileMotionNeck;
-                if (anim.FileMotionNeck != null)
-                    donorInfo.paramFemale.fileMotionNeck = anim.FileMotionNeck;
-                if (anim.IsFemaleInitiative != null)
-                    donorInfo.isFemaleInitiative = anim.IsFemaleInitiative.Value;
-                if (anim.FileSiruPaste != null && SiruPasteFiles.TryGetValue(anim.FileSiruPaste.ToLower(), out var fileSiruPaste))
-                    donorInfo.paramFemale.fileSiruPaste = fileSiruPaste;
-
-                donorInfo.lstCategory = anim.categories.Select(c =>
-                    {
-                        var cat = new HSceneProc.Category();
-                        cat.category = (int)c;
-                        return cat;
-                    }
-                ).ToList();
-
-                Logger.LogDebug("Adding anim " + anim.AnimationName + " to EMode " + anim.Mode);
-                animListInfo.Add(donorInfo);
-
-                swapAnimationMapping[donorInfo] = anim;
-            }
-        }
-
-        [HarmonyTranspiler, HarmonyPatch(typeof(HSprite), nameof(HSprite.OnChangePlaySelect))]
-        private static IEnumerable<CodeInstruction> OnChangePlaySelect(IEnumerable<CodeInstruction> instructions)
-        {
-            //force position change even if position appears to match. Prevents clicks from being eaten.
-            return new CodeMatcher(instructions)
-                .MatchForward(false, 
-                    new CodeMatch(OpCodes.Ldfld, AccessTools.Field(typeof(HSprite), "flags")),
-                    new CodeMatch(OpCodes.Ldfld, AccessTools.Field(typeof(HFlag), "nowAnimationInfo")),
-                    new CodeMatch(OpCodes.Ldfld, AccessTools.Field(typeof(HSceneProc.AnimationListInfo), "id"))
-                )
-                .MatchForward(false,
-                    new CodeMatch(OpCodes.Ret)
-                )
-                .SetAndAdvance(OpCodes.Nop, null)
-                .InstructionEnumeration();
-        }
-
-        [HarmonyPostfix, HarmonyPatch(typeof(HSprite), nameof(HSprite.LoadMotionList))]
-        private static void LoadMotionList(HSprite __instance, List<HSceneProc.AnimationListInfo> _lstAnimInfo, GameObject _objParent)
-        {
-            if(_lstAnimInfo == null || _lstAnimInfo.Count == 0)
-                return;
-
-            
-            var buttonParent = _objParent.transform;
-            Transform scrollT = null;
-            if(vrType != null || UseGrid.Value)
-            {
-                DestroyImmediate(_objParent.GetComponent<VerticalLayoutGroup>());
-                DestroyImmediate(_objParent.GetComponent<GridLayoutGroup>());
-                DestroyImmediate(_objParent.GetComponent<ContentSizeFitter>());
-                var glg = _objParent.AddComponent<GridLayoutGroup>();
-                glg.cellSize = new Vector2(200, 35);
-                glg.startAxis = GridLayoutGroup.Axis.Vertical;
-                glg.startCorner = GridLayoutGroup.Corner.UpperRight;
-                glg.constraint = GridLayoutGroup.Constraint.FixedRowCount;
-                glg.constraintCount = 15;
-                glg.childAlignment = TextAnchor.UpperRight;
-            }
-            else
-            {
-                var buttons = _objParent.transform.Cast<Transform>().ToList();
-                
-                var go = DefaultControls.CreateScrollView(new DefaultControls.Resources());
-                go.transform.SetParent(_objParent.transform, false);
-                var scroll = go.GetComponent<ScrollRect>();
-                scroll.horizontal = false;
-                scroll.scrollSensitivity = 32f;
-                scroll.movementType = ScrollRect.MovementType.Clamped;
-                scroll.horizontalScrollbarVisibility = ScrollRect.ScrollbarVisibility.AutoHide;
-                scroll.verticalScrollbarVisibility = ScrollRect.ScrollbarVisibility.AutoHide;
-                DestroyImmediate(scroll.horizontalScrollbar.gameObject);
-                DestroyImmediate(scroll.verticalScrollbar.gameObject);
-                DestroyImmediate(scroll.GetComponent<Image>());
-
-                var copyTarget = GameObject.Find("Canvas").transform.Find("clothesFileWindow/Window/WinRect/ListArea/Scroll View/Scrollbar Vertical").gameObject;
-                var newScrollbar = Instantiate(copyTarget, go.transform);
-                scroll.verticalScrollbar = newScrollbar.GetComponent<Scrollbar>();
-                newScrollbar.transform.SetRect(1f, 0f, 1f, 1f, 0f, 0f, 18f);
-                
-                var triggerEvent = new EventTrigger.TriggerEvent();
-                triggerEvent.AddListener(x => GlobalMethod.SetCameraMoveFlag(__instance.flags.ctrlCamera, false));
-                var eventTrigger = newScrollbar.AddComponent<EventTrigger>();
-                eventTrigger.triggers.Add(new EventTrigger.Entry{eventID = EventTriggerType.PointerDown, callback = triggerEvent });
-
-                var vlg = _objParent.GetComponent<VerticalLayoutGroup>();
-                var csf = _objParent.GetComponent<ContentSizeFitter>();
-                vlg.enabled = false;
-                csf.enabled = false;
-                CopyComponent(vlg, scroll.content.gameObject).enabled = true;
-                CopyComponent(csf, scroll.content.gameObject).enabled = true;
-            
-                // remove the buttons as we're going to rebuild the entire list
-                buttons.ForEach(x => Destroy(x.gameObject));
-
-                buttonParent = scroll.content;
-                scrollT = scroll.gameObject.transform;
-            }
-            
-            foreach(var anim in _lstAnimInfo)
-            {
-                var btn = Instantiate(__instance.objMotionListNode, buttonParent, false);
-                btn.AddComponent<HSprite.AnimationInfoComponent>().info = anim;
-                var label = btn.GetComponentInChildren<TextMeshProUGUI>();
-                label.text = anim.nameAnimation;
-                label.color = Color.black;
-                
-                //TODO: wat
-                var tgl = btn.GetComponent<Toggle>();
-                tgl.group = _objParent.GetComponent<ToggleGroup>();
-                tgl.enabled = false;
-                tgl.enabled = true;
-
-                btn.GetComponent<SceneAssist.PointerAction>().listClickAction.Add(() =>
-                {
-                    __instance.OnChangePlaySelect(btn);
-                });
-
-                btn.SetActive(true);
-                if(__instance.flags.nowAnimationInfo == anim)
-                    btn.GetComponent<Toggle>().isOn = true;
-
-                swapAnimationMapping.TryGetValue(anim, out var swap);
-                if (swap != null)
-                {
-                    btn.transform.FindLoop("Background").GetComponent<Image>().color = buttonColor;
-                    label.text = swap.AnimationName;
-                }
-            }
-
-            // order all buttons by name
-            var allButtons = buttonParent.Cast<Transform>().OrderBy(x => x.GetComponentInChildren<TextMeshProUGUI>().text).ToList();
-            foreach(var t in allButtons)
-            {
-                // disable New text
-                var newT = t.FindLoop("New");
-                if(newT) newT.gameObject.SetActive(false);
-                
-                if(SortPositions.Value)
-                    t.SetAsLastSibling();
-
-                var textMeshGo = t.FindLoop("TextMeshPro Text").gameObject;
-                var textMesh = textMeshGo.GetComponent<TextMeshProUGUI>();
-                textMesh.enableWordWrapping = false;
-                textMesh.overflowMode = TextOverflowModes.Overflow; // disable ... after text
-
-                // add scrolling text if text is long enough
-                var rectT = (RectTransform)t;
-                if(rectT.sizeDelta.x < textMesh.preferredWidth)
-                {
-                    textMesh.alignment = TextAlignmentOptions.CaplineLeft;
-
-                    var txtScroll = textMeshGo.AddComponent<TextScroll>();
-                    txtScroll.textMesh = textMesh;
-                    txtScroll.transBase = rectT;
-                }
-            }
-
-            if(scrollT != null && allButtons.Count > 8)
-                scrollT.SetRect(0f, 0f, 1f, 1f, -5f, -100f, -5f, 100f);
-        }
-
-        [HarmonyPostfix, HarmonyPatch(typeof(HSceneProc), nameof(HSceneProc.ChangeAnimator))]
-        private static void SwapAnimation(object __instance, HSceneProc.AnimationListInfo _nextAinmInfo)
-        {
-            if (!swapAnimationMapping.TryGetValue(_nextAinmInfo, out var swapAnimationInfo))
-            {
-                return;
-            }
-
-            RuntimeAnimatorController femaleCtrl = null;
-            RuntimeAnimatorController maleCtrl = null;
-            if(!string.IsNullOrEmpty(swapAnimationInfo.PathFemale) || !string.IsNullOrEmpty(swapAnimationInfo.ControllerFemale))
-                femaleCtrl = AssetBundleManager.LoadAsset(swapAnimationInfo.PathFemale, swapAnimationInfo.ControllerFemale, typeof(RuntimeAnimatorController)).GetAsset<RuntimeAnimatorController>();
-            if(!string.IsNullOrEmpty(swapAnimationInfo.PathMale) || !string.IsNullOrEmpty(swapAnimationInfo.ControllerMale))
-                maleCtrl = AssetBundleManager.LoadAsset(swapAnimationInfo.PathMale, swapAnimationInfo.ControllerMale, typeof(RuntimeAnimatorController)).GetAsset<RuntimeAnimatorController>();
-            
-            var t_hsp = Traverse.Create(__instance);
-            var female = t_hsp.Field<List<ChaControl>>("lstFemale").Value[0];
-            var male = t_hsp.Field<ChaControl>("male").Value;
-            ////TODO: lstFemale[1], male1
-
-            if(femaleCtrl != null)
-                female.animBody.runtimeAnimatorController = SetupAnimatorOverrideController(female.animBody.runtimeAnimatorController, femaleCtrl);
-            if(maleCtrl != null)
-                male.animBody.runtimeAnimatorController = SetupAnimatorOverrideController(male.animBody.runtimeAnimatorController, maleCtrl);
-
-            var mi = t_hsp.Field<List<MotionIK>>("lstMotionIK").Value;
-            mi.ForEach(mik => mik.Release());
-            mi.Clear();
-
-            //TODO: MotionIKData.
-            mi.Add(new MotionIK(female));
-            mi.Add(new MotionIK(male));
-            mi.ForEach(mik =>
-            {
-                mik.SetPartners(mi);
-                mik.Reset();
-            });
-        }
-
         private static AnimatorOverrideController SetupAnimatorOverrideController(RuntimeAnimatorController src, RuntimeAnimatorController over)
         {
             if(src == null || over == null)
@@ -407,75 +175,5 @@ namespace AnimationLoader
             return dst;
         }
 
-        [HarmonyPostfix, HarmonyPatch(typeof(Studio.Info), nameof(Studio.Info.LoadExcelDataCoroutine))]
-        private static void LoadStudioAnims(Studio.Info __instance, ref IEnumerator __result)
-        {
-            __result = __result.AppendCo(() =>
-            {
-                foreach(var keyVal in animationDict)
-                {
-                    CreateGroup(0);
-                    CreateGroup(1);
-
-                    void CreateGroup(byte sex)
-                    {
-                        var grp = new Info.GroupInfo{ name = $"AL {(sex == 0 ? "M" : "F")} {keyVal.Key}" };
-                        var animGrp = new Dictionary<int, Dictionary<int, Info.AnimeLoadInfo>>();
-
-                        var grpKey = $"{keyVal.Key}{sex}";
-                        if(!EModeGroups.TryGetValue(grpKey, out var grpId))
-                            return;
-            
-                        foreach(var swapAnimInfo in keyVal.Value.Where(x => x.StudioId >= 0))
-                        {
-                            var path = sex == 0 ? swapAnimInfo.PathMale : swapAnimInfo.PathFemale;
-                            var ctrl = sex == 0 ? swapAnimInfo.ControllerMale : swapAnimInfo.ControllerFemale;
-
-                            if(string.IsNullOrEmpty(path) || string.IsNullOrEmpty(ctrl))
-                                continue;
-                            
-                            var controller = AssetBundleManager.LoadAsset(path, ctrl, typeof(RuntimeAnimatorController)).GetAsset<RuntimeAnimatorController>();
-                            if(controller == null)
-                                continue;
-                            
-                            var animName = string.IsNullOrEmpty(swapAnimInfo.AnimationName) ? ctrl : swapAnimInfo.AnimationName;
-                            grp.dicCategory.Add(swapAnimInfo.StudioId, animName);
-                            var animCat = new Dictionary<int, Info.AnimeLoadInfo>();
-                            animGrp.Add(swapAnimInfo.StudioId, animCat);
-                            
-                            var clips = controller.animationClips;
-                            for(int i = 0; i < clips.Length; i++)
-                            {
-                                var newSlot = UniversalAutoResolver.GetUniqueSlotID();
-
-                                UniversalAutoResolver.LoadedStudioResolutionInfo.Add(new StudioResolveInfo
-                                {
-                                    GUID = swapAnimInfo.Guid,
-                                    Slot = i,
-                                    ResolveItem = true,
-                                    LocalSlot = newSlot,
-                                    Group = grpId,
-                                    Category = swapAnimInfo.StudioId
-                                });
-                                
-                                animCat.Add(newSlot, new Info.AnimeLoadInfo
-                                {
-                                    name = clips[i].name,
-                                    bundlePath = path,
-                                    fileName = ctrl,
-                                    clip = clips[i].name,
-                                });
-                            }
-                        }
-
-                        if(animGrp.Count > 0)
-                        {
-                            __instance.dicAGroupCategory.Add(grpId, grp);
-                            __instance.dicAnimeLoadInfo.Add(grpId, animGrp);
-                        }
-                    }
-                }
-            });
-        }
     }
 }
